+++ core/src/main/scala/jdksymlink/core/data.scala
@@ -1,4 +1,4 @@
-package jdksymlink
+package jdksymlink.core
 
 import java.io.File
